Improve decoding JSON
The `Json` class from the crwlr/utils is now used to decode JSON
strings. It tries to fix keys without quotes, which is allowed in
relaxed JSON. Further, JSON-LD <script> blocks containing an invalid
JSON string are ignored and don't lead to an error anymore.

Also, you can now pass a PSR-3 LoggerInterface to the `SchemaOrg` class,
so it'll log decoding errors.