Refactoring
As some methods in the `SchemaOrg` class got pretty long and also there
were more and more methods dealing with arrays (the JSON converted to
array), refactor a bit. Create a new class DataArray for dealing with
the arrays and split up some methods.
Also add more tests and add the changelog entry for the new feature.